# meta/src/main/resources/docs/swagger.yml
---
openapi: 3.0.0
info:
  version: 0.0.1
  title: LaserRabbit
  description: |
    Is storing data in git viable?

servers:
  - url: http://localhost:6969

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Form:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string

      required:
        - firstName
        - lastName

paths:
  /forms:
    get:
      security:
        - bearerAuth: []
      description: Returns all forms
      operationId: allForms
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Form'
        401:
          description: JWT probably missing / invalid

    post:
      security:
        - bearerAuth: []
      description: Create a new form
      operationId: createForm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Form'
      responses:
        200:
          description: Effective result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        303:
          description: The url of the new form
        401:
          description: JWT probably missing / invalid

  /form/{formId}:
    get:
      security:
        - bearerAuth: []
      description: Reads this version of a form
      operationId: readForm
      responses:
        200:
          description: Form found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        404:
          description: Form missing
        401:
          description: JWT probably missing / invalid

      parameters:
        - name: formId
          in: path
          schema:
            type: string
          required: true